refactor: 채팅 로직 패키지 구조 리팩토링
