--- src/main/java/com/becareful/becarefulserver/domain/caregiver/service/CaregiverChatService.java
@@ -1,40 +1,0 @@
-package com.becareful.becarefulserver.domain.caregiver.service;
-
-import com.becareful.becarefulserver.domain.caregiver.domain.Caregiver;
-import com.becareful.becarefulserver.domain.caregiver.dto.response.ChatroomResponse;
-import com.becareful.becarefulserver.domain.matching.domain.Matching;
-import com.becareful.becarefulserver.domain.matching.repository.CompletedMatchingRepository;
-import com.becareful.becarefulserver.domain.matching.repository.ContractRepository;
-import com.becareful.becarefulserver.domain.matching.repository.MatchingRepository;
-import com.becareful.becarefulserver.global.util.AuthUtil;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class CaregiverChatService {
-
-    private final AuthUtil authUtil;
-    private final MatchingRepository matchingRepository;
-    private final ContractRepository contractRepository;
-    private final CompletedMatchingRepository completedMatchingRepository;
-
-    @Transactional(readOnly = true)
-    public List<ChatroomResponse> getChatList() {
-        Caregiver caregiver = authUtil.getLoggedInCaregiver();
-        List<Matching> matchingList = matchingRepository.findByCaregiver(caregiver);
-
-        List<ChatroomResponse> responses = new ArrayList<>();
-        matchingList.forEach(matching -> {
-            contractRepository.findTop1ByMatchingOrderByCreateDateDesc(matching).ifPresent(contract -> {
-                boolean isCompleted = completedMatchingRepository.existsCompletedMatchingByContract(contract);
-                ChatroomResponse.of(matching, contract, isCompleted);
-            });
-        });
-
-        return responses;
-    }
-}
