+++ src/main/java/com/becareful/becarefulserver/domain/chat/service/CaregiverChatService.java
@@ -1,0 +1,71 @@
+package com.becareful.becarefulserver.domain.chat.service;
+
+import static com.becareful.becarefulserver.global.exception.ErrorMessage.CONTRACT_NOT_EXISTS;
+import static com.becareful.becarefulserver.global.exception.ErrorMessage.MATCHING_NOT_EXISTS;
+
+import com.becareful.becarefulserver.domain.caregiver.domain.Caregiver;
+import com.becareful.becarefulserver.domain.chat.dto.ChatroomResponse;
+import com.becareful.becarefulserver.domain.matching.domain.CompletedMatching;
+import com.becareful.becarefulserver.domain.matching.domain.Contract;
+import com.becareful.becarefulserver.domain.matching.domain.Matching;
+import com.becareful.becarefulserver.domain.matching.dto.response.ContractInfoListResponse;
+import com.becareful.becarefulserver.domain.matching.repository.CompletedMatchingRepository;
+import com.becareful.becarefulserver.domain.matching.repository.ContractRepository;
+import com.becareful.becarefulserver.domain.matching.repository.MatchingRepository;
+import com.becareful.becarefulserver.global.exception.exception.ContractException;
+import com.becareful.becarefulserver.global.exception.exception.MatchingException;
+import com.becareful.becarefulserver.global.util.AuthUtil;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class CaregiverChatService {
+
+    private final AuthUtil authUtil;
+    private final MatchingRepository matchingRepository;
+    private final ContractRepository contractRepository;
+    private final CompletedMatchingRepository completedMatchingRepository;
+
+    @Transactional(readOnly = true)
+    public List<ChatroomResponse> getChatList() {
+        Caregiver caregiver = authUtil.getLoggedInCaregiver();
+        List<Matching> matchingList = matchingRepository.findByCaregiver(caregiver);
+
+        List<ChatroomResponse> responses = new ArrayList<>();
+        matchingList.forEach(matching -> {
+            contractRepository.findTop1ByMatchingOrderByCreateDateDesc(matching).ifPresent(contract -> {
+                boolean isCompleted = completedMatchingRepository.existsCompletedMatchingByContract(contract);
+                ChatroomResponse.of(matching, contract, isCompleted);
+            });
+        });
+
+        return responses;
+    }
+
+    @Transactional(readOnly = true)
+    public ContractInfoListResponse getChatRoomDetailData(Long matchingId) {
+        Caregiver caregiver = authUtil.getLoggedInCaregiver();
+
+        List<Contract> contracts = contractRepository.findByMatchingIdOrderByCreateDateAsc(matchingId);
+        Matching matching =
+                matchingRepository.findById(matchingId).orElseThrow(() -> new MatchingException(MATCHING_NOT_EXISTS));
+
+        matching.validateCaregiver(caregiver.getId());
+
+        return ContractInfoListResponse.of(matching, contracts);
+    }
+
+    @Transactional
+    public void createCompletedMatching(Long contractId) {
+        Caregiver loggedInCaregiver = authUtil.getLoggedInCaregiver();
+        Contract contract =
+                contractRepository.findById(contractId).orElseThrow(() -> new ContractException(CONTRACT_NOT_EXISTS));
+
+        CompletedMatching completedMatching = new CompletedMatching(loggedInCaregiver, contract);
+        completedMatchingRepository.save(completedMatching);
+    }
+}
