+++ src/main/java/com/becareful/becarefulserver/domain/matching/domain/Matching.java
@@ -1,6 +1,6 @@
 package com.becareful.becarefulserver.domain.matching.domain;
 
-import static com.becareful.becarefulserver.global.exception.ErrorMessage.MATCHING_CANNOT_REJECT;
+import static com.becareful.becarefulserver.global.exception.ErrorMessage.*;
 
 import com.becareful.becarefulserver.domain.caregiver.domain.WorkApplication;
 import com.becareful.becarefulserver.domain.common.domain.BaseEntity;
@@ -8,6 +8,7 @@ import com.becareful.becarefulserver.domain.matching.domain.converter.MediationT
 import com.becareful.becarefulserver.domain.matching.domain.vo.MatchingResultInfo;
 import com.becareful.becarefulserver.domain.matching.domain.vo.MatchingResultStatus;
 import com.becareful.becarefulserver.domain.matching.dto.request.RecruitmentMediateRequest;
+import com.becareful.becarefulserver.global.exception.exception.MatchingException;
 import com.becareful.becarefulserver.global.exception.exception.RecruitmentException;
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.AttributeOverrides;
@@ -161,14 +162,30 @@ public class Matching extends BaseEntity {
     }
 
     private void validateMatchingCompletable() {
-        if (!this.matchingApplicationStatus.equals(MatchingApplicationStatus.지원검토중)) {
-            throw new RecruitmentException("지원한 경우에만 합격, 불합격 처리할 수 있습니다.");
+        if (matchingApplicationStatus.equals(MatchingApplicationStatus.지원검토중)) {
+            return;
         }
+        throw new RecruitmentException("지원한 경우에만 합격, 불합격 처리할 수 있습니다.");
     }
 
     private void validateMatchingUpdatable() {
-        if (!this.matchingApplicationStatus.equals(MatchingApplicationStatus.미지원)) {
-            throw new RecruitmentException(MATCHING_CANNOT_REJECT);
+        if (matchingApplicationStatus.equals(MatchingApplicationStatus.미지원)) {
+            return;
         }
+        throw new RecruitmentException(MATCHING_CANNOT_REJECT);
+    }
+
+    public void validateCaregiver(Long caregiverId) {
+        if (workApplication.getCaregiver().getId().equals(caregiverId)) {
+            return;
+        }
+        throw new MatchingException(MATCHING_CAREGIVER_DIFFERENT);
+    }
+
+    public void validateSocialWorker(Long socialWorkerId) {
+        if (workApplication.getCaregiver().getId().equals(socialWorkerId)) {
+            return;
+        }
+        throw new MatchingException(MATCHING_SOCIAL_WORKER_DIFFERENT);
     }
 }
