+++ src/main/java/com/becareful/becarefulserver/domain/matching/service/CompletedMatchingService.java
@@ -2,11 +2,9 @@ package com.becareful.becarefulserver.domain.matching.service;
 
 import com.becareful.becarefulserver.domain.caregiver.domain.Caregiver;
 import com.becareful.becarefulserver.domain.matching.domain.CompletedMatching;
-import com.becareful.becarefulserver.domain.matching.domain.Contract;
 import com.becareful.becarefulserver.domain.matching.dto.request.EditCompletedMatchingNoteRequest;
 import com.becareful.becarefulserver.domain.matching.dto.response.CompletedMatchingInfoResponse;
 import com.becareful.becarefulserver.domain.matching.repository.CompletedMatchingRepository;
-import com.becareful.becarefulserver.domain.matching.repository.ContractRepository;
 import com.becareful.becarefulserver.global.util.AuthUtil;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +18,6 @@ public class CompletedMatchingService {
 
     private final CompletedMatchingRepository completedMatchingRepository;
     private final AuthUtil authUtil;
-    private final ContractRepository contractRepository;
 
     public List<CompletedMatchingInfoResponse> getCompletedMatchings() {
         Caregiver caregiver = authUtil.getLoggedInCaregiver();
@@ -36,16 +33,5 @@ public class CompletedMatchingService {
                 .orElseThrow(() -> new IllegalArgumentException("Matching not found"));
 
         completedMatching.updateNote(request.note());
-    }
-
-    @Transactional
-    public void createCompletedMatching(Long contractId) {
-        Caregiver loggedInCaregiver = authUtil.getLoggedInCaregiver();
-        Contract contract = contractRepository
-                .findById(contractId)
-                .orElseThrow(() -> new IllegalArgumentException("Contract not found"));
-
-        CompletedMatching completedMatching = new CompletedMatching(loggedInCaregiver, contract);
-        completedMatchingRepository.save(completedMatching);
     }
 }
