+++ src/main/java/com/becareful/becarefulserver/domain/socialworker/service/SocialWorkerService.java
@@ -2,14 +2,10 @@ package com.becareful.becarefulserver.domain.socialworker.service;
 
 import static com.becareful.becarefulserver.global.exception.ErrorMessage.*;
 
-import com.becareful.becarefulserver.domain.caregiver.domain.Caregiver;
 import com.becareful.becarefulserver.domain.common.vo.Gender;
-import com.becareful.becarefulserver.domain.matching.domain.Contract;
 import com.becareful.becarefulserver.domain.matching.domain.Matching;
 import com.becareful.becarefulserver.domain.matching.domain.Recruitment;
 import com.becareful.becarefulserver.domain.matching.dto.ElderlySimpleDto;
-import com.becareful.becarefulserver.domain.matching.repository.CompletedMatchingRepository;
-import com.becareful.becarefulserver.domain.matching.repository.ContractRepository;
 import com.becareful.becarefulserver.domain.matching.repository.MatchingRepository;
 import com.becareful.becarefulserver.domain.nursing_institution.domain.NursingInstitution;
 import com.becareful.becarefulserver.domain.nursing_institution.repository.NursingInstitutionRepository;
@@ -20,7 +16,6 @@ import com.becareful.becarefulserver.domain.socialworker.domain.vo.AssociationRa
 import com.becareful.becarefulserver.domain.socialworker.dto.SocialWorkerSimpleDto;
 import com.becareful.becarefulserver.domain.socialworker.dto.request.SocialWorkerCreateRequest;
 import com.becareful.becarefulserver.domain.socialworker.dto.request.SocialWorkerUpdateBasicInfoRequest;
-import com.becareful.becarefulserver.domain.socialworker.dto.response.ChatList;
 import com.becareful.becarefulserver.domain.socialworker.dto.response.SocialWorkerHomeResponse;
 import com.becareful.becarefulserver.domain.socialworker.dto.response.SocialWorkerMyResponse;
 import com.becareful.becarefulserver.domain.socialworker.repository.ElderlyRepository;
@@ -34,14 +29,12 @@ import com.becareful.becarefulserver.global.util.JwtUtil;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -58,8 +51,6 @@ public class SocialWorkerService {
     private final NursingInstitutionRepository nursingInstitutionRepository;
 
     private final MatchingRepository matchingRepository;
-    private final ContractRepository contractRepository;
-    private final CompletedMatchingRepository completedMatchingRepository;
     private final ElderlyRepository elderlyRepository;
 
     private final AuthUtil authUtil;
@@ -177,84 +168,6 @@ public class SocialWorkerService {
     public SocialWorkerMyResponse getMyInfo() {
         SocialWorker loggedInSocialWorker = authUtil.getLoggedInSocialWorker();
         return SocialWorkerMyResponse.from(loggedInSocialWorker);
-    }
-
-    @Transactional
-    public ChatList getChatList() {
-        SocialWorker socialworker = authUtil.getLoggedInSocialWorker();
-        NursingInstitution nursingInstitution = socialworker.getNursingInstitution();
-        List<Matching> matchingList = matchingRepository.findByNursingInstitution(nursingInstitution);
-
-        List<ChatList.ChatroomInfo> chatroomInfoList = matchingList.stream()
-                .map(matching -> {
-                    // Elderly와 Caregiver 정보를 가져오기
-                    Caregiver caregiver = matching.getWorkApplication().getCaregiver();
-                    Elderly elderly = matching.getRecruitment().getElderly();
-                    String timeDifference = getTimeDifferenceString(matching);
-                    Contract latestContract = contractRepository
-                            .findTop1ByMatchingOrderByCreateDateDesc(matching)
-                            .get();
-                    String recentChat = isContractInCompletedMatching(matching) ? "최종 승인이 확정되었습니다!" : "합격 축하드립니다.";
-
-                    // ChatroomInfo 생성
-                    return new ChatList.ChatroomInfo(
-                            matching.getId(),
-                            caregiver.getProfileImageUrl(), // 어르신 프로필 이미지 URL
-                            caregiver.getName(), // 요양보호자 이름
-                            recentChat, // 최근 채팅
-                            timeDifference,
-                            elderly.getName(), // 어르신 이름
-                            elderly.getAge(), // 어르신 나이
-                            elderly.getGender() // 어르신 성별
-                            );
-                })
-                .collect(Collectors.toList());
-
-        // ChatList 반환
-        return new ChatList(chatroomInfoList);
-    }
-
-    public LocalDateTime findLatestContractCreatedDate(Matching matching) {
-        Contract latestContract = contractRepository
-                .findTop1ByMatchingOrderByCreateDateDesc(matching)
-                .orElse(null);
-        return latestContract != null ? latestContract.getCreateDate() : null;
-    }
-
-    public boolean isContractInCompletedMatching(Matching matching) {
-        Contract latestContract = contractRepository
-                .findTop1ByMatchingOrderByCreateDateDesc(matching)
-                .orElse(null);
-        if (latestContract != null) {
-            return completedMatchingRepository.existsCompletedMatchingByContract(latestContract);
-        }
-        return false;
-    }
-
-    public String getTimeDifferenceString(Matching matching) {
-        // 현재 시간
-        LocalDateTime currentTime = LocalDateTime.now();
-
-        // 가장 최신 Contract의 생성 시간
-        LocalDateTime contractCreatedTime = findLatestContractCreatedDate(matching);
-
-        // Duration을 사용하여 차이 계산
-        Duration duration = Duration.between(contractCreatedTime, currentTime);
-
-        // 차이에 따라 다른 시간 단위로 변환
-        if (duration.toHours() < 1) {
-            // 1시간 이내이면 분 단위로 반환
-            long minutes = duration.toMinutes();
-            return minutes + "분 전";
-        } else if (duration.toDays() < 1) {
-            // 1일 이내이면 시간 단위로 반환
-            long hours = duration.toHours();
-            return hours + "시간 전";
-        } else {
-            // 1일 이상이면 일 단위로 반환
-            long days = duration.toDays();
-            return days + "일 전";
-        }
     }
 
     private void validateEssentialAgreement(boolean isAgreedToTerms, boolean isAgreedToCollectPersonalInfo) {
