+++ src/main/java/com/becareful/becarefulserver/global/exception/exception/ContractException.java
@@ -1,0 +1,7 @@
+package com.becareful.becarefulserver.global.exception.exception;
+
+public class ContractException extends DomainException {
+    public ContractException(String message) {
+        super(message);
+    }
+}
