+++ src/main/java/com/becareful/becarefulserver/global/exception/exception/MatchingException.java
@@ -1,0 +1,7 @@
+package com.becareful.becarefulserver.global.exception.exception;
+
+public class MatchingException extends DomainException {
+    public MatchingException(String message) {
+        super(message);
+    }
+}
